Fixes, enhancements
Changed the unique ID for blocking filter to be based on the machine
name.
Fixed synchronization issue with WinSparkle shutdown - need to promote
the shutdown call in the callback to the primary application thread.
Added overload extension in Application to take a custom enum for
shutdown reasons. This way we can shutdown with a given reason from
anywhere, and our handler for the shutdown event with App.xaml.cs will
get this reason in a central location, where we can then modify our
shutdown actions.
Fixed som synchronization stuff with the terms view model, was pegging
the CPU requesting the terms string over and over.
In WebServiceUtils, where the bool to set no logging of exceptions is
present, actually acted on this and will not log exceptions if its set
appropriately.