Fixes
Fixes #5.
Attempt to address #6 by writing user auth data to a random file, then
performing a move or overwrite of the file on save. This has not been
effective when the task scheduler is responsible for force terminating
the process, but the changes are committed nontheless.
Modified StringExtensions.Valid() to include a check on the length of
the string.
Removed old Process invoking based task scheduler interaction and
replaced it with the task scheduler library. This way we can set the
priority of the startup item to maximum.
Isolated various code in the shutdown event handlers in their own
try/catch blocks to ensure maximum fault tolerance. This ensures that
critical code later on in the function scope is not abandoned because
other code in the same block failed.