Fixes, enhancements
Fixes an issue where the relaxed policy counters and such were being
messed up by subscribing to the same events with the same handlers
multiple times.
Added an installer class that enables the app to be launched during the
installation process.
Modified the installer projects to launch the application when the
installation is committed.
Improved the fast html text extractrator by making it only copy letter,
digit and punctuation characters. This avoids copying in junk.
Bump version to 0.4.15.