Custom written single instance enforcement
Custom single instance enforcement. The one provided in .NET doesn't work globally, or I'm missing something I dunno. It didn't work properly AFAIK and we would sometimes get two instances still.

This version functions by creating a global mutext. Further, we hook window functions in our base window object and broadcast an event when the window should be restored (when we get various ShowWindow commands), so that the main application knows to bring itself out of the task tray when this happens.

When we could not acquire the global mutex, we find the process by name and then enumerate all child widows of the current running process and spam them with show window messages. This should trigger the aforementioned code to fire and show the current running instance, before the new instance simply exits.

One noted issue is that when the post install command runs from the installer to start the application, this functionality doesn't work. It would appear that the MSI window is soaking up these messages and preventing them from hitting the actual app.

Bump version to 0.6.27.