Move to multiprocess
Warning: Commit is broken. Installer projects are non-functional and need to be re-created.
Move to multiprocess.
Filtering is now done in a SYSTEM process and communications bidirectionally with GUI clients over IPC.
GUI is now independent of the filtering itself.
New service dynamically compiles and emits services that will protect each other. They all monitor eachother and restart one another if any one of them is terminated.
Removed provider conditions view. This is not needed, the EULA can be presented in the installer. Unnecessary complexity.