Updates and fixes
Restructuring of the common utilities files to organize into folders so it's not such a mess. Namespaces left unchanged though.

Isolated web service utility to be used solely by the filtering service. It was a leaky mess having it be a public shared class.

Copied internet availability check directly into the main window so it can get the status independent of the web service util.

Added last updated timestamp to settings view.

Removed the manifest for the GUI so it no longer needs to run as admin. Because we have removed the manifest, we can launch the GUI from a service by impersonating the currently logged on user.

Added code to the GUI app to detect if the service is running. If not, the program will restart itself as admin, and force the services to start.

Stripped down the web service util considerably to simplify it. We longer use or care about CSRF tokens. We no longer care about re-authing. We use long lived access tokens instead. These tokens are stored in the registry on success.

Because we removed the manifest from the GUI, and can start the GUI as the current user non-elevated, the filtering service will force the GUI to run if and when auth fails and immediately ask the user to re-auth.

Added a simple flag for when auth fails to prevent filtering. That is, no rules will change, the filter service will not be stooped nor rules unloaded. The service will continue to operate in its loop (update checks etc), the only thing that will happen is that the callbacks for inspecting requests will immediately return when not authorized.

Renamed process protection class to CriticalKernelProcessUtility.

Added external code under contrib that uses WinAPI to launch non-manifest applications from a service.

Update to latest HttpFilteringEngine, which solves problems with encryption/incomplete certificate chains.

Unblock request is temporarily disabled because of changes to isolate the service utility. The simplest thing would be IPC. Perhaps we can do this because of our new class to run things as the current logged in user.

Fixes #32
Fixes #33
Fixes #28
Fixes #26
Fixes #25
Fixes #14
Fixes #5