Fixes
App and app library versions now bumped above default value of 1.0.0.0 to enforce installation of updated binaries. If your binary versions are less than the current installed version, the windows installer will ignore your "newer" versions.

Made the service force-start the GUI whenever it starts up, as well as whenever a user logs in. So, every user will get a GUI instance run when they log on regardless and so long as the app is installed, the GUI will always run at startup minimized and you can't change or stop this.

Coded the service and GUI to both fully shut down when auth is required and the user tries to close the app. So, if the main window close button was clicked and we were on the login screen, the GUI will shut down. When a GUI shuts down, the server gets an IPC client disconnect message and we decrement the amount of currently connected clients. If the service is awaiting auth, and the count goes to or below zero, the service will entirely shut down without safeguards.

Fixes #52
Fixes #38
Fixes #27