Upgrade to new engine, restructure projects for Kestrel/LibUV
Upgraded to use latest engine.
Created arch-specific project files for the engine and the GUI. This makes it so we can debug easily and also is necessary because of arch-specific LibUV package resolution via nuget. The arch needs to be set specifically for Kestrel to grab the correct libUV dll.
Switch to VS2017.
Switch Nuget management to Package Reference.