Remove System.Data.Sqlite, other fixes
Fixes #78
Removes System.Data.Sqlite and now uses Microsoft.Data.Sqlite
Installs the Microsoft.Data.Sqlite.Core package as well, even though it seems to be identical to Microsoft.Data.Sqlite. By doing this, the nuget weirdness seems to have been resolved.
Removed ca-bundle as a built in resource for the service as it's no longer needed.
Removed code that extracts and appends stuff to ca-bundle.
Removed old firefox trust code from the GUI project.
Removed old code for purging the cache for System.Data.Sqlite in-memory databases. This should no longer be necessary with the Microsoft.Data.Sqlite bundle, but keep an eye on it.